Started using PEP517/518 for builds
It took my a while to figure out how to get dependencies installed
before running any of the functions in setup.py. Turns out, there
was a pep for just that. It defines a `pyproject.toml` file and adds
build dependencies there. The subprocess for installaing packages
during `setup.py` has been removed.