Homogenized C++ and Python api docs
I hate frontend... but that was the easiest way to fix this.
Look into the docs-source/source/_static/add_class.js file.
It seems breathe and sphinx don't add the same HTML classes.
This might be specific to using the RTD theme.
I fixed this by injecting and remocing the appropriate classes.